More Changes to Tile System
Had to revert back to a single image for the background it was a bit
unstable for the prototype a single image is fine in the future I'll
have to refactor that. I also implemented a crude encounter mechanic.
The Art still needs to be updated but that's a simple thing, mostly time
consuming.