Base Game Director Script
Created a base game director script, I also added an audio and input
manager as a test they have no functionality. There are a couple of
issues with the approach we discussed at the meeting, it shouldn't
affect us much but nevertheless I left some comments in the code.
Remember from now on branch off from Test and work on the different
scripts when you're done merge it back in and make sure the the scripts
integrate correctly.