Input Manager
Some progress with the input manager, I added a timer class that I may
turn into a timer manager class so we can keep different timers for
other parts of the game. The input now has a sequential list that resets
after 3 second of inactivity.

The other part to deal with now is creating a command pattern for inputs
so that swaping them around is going to be quite easy in the future.