Completed Basic Audio Manager
Corrected a small bug with the hud manager in case of a new scene.
completed a basic audio manager, started to work a finalized version of
the QTE mechanic. Still need to change the way the list of inputs is
cleared and do a second pass on the timer and switch to coroutines.
Lastly I need to clean up the directory files from the latest level
build and add it back to this solution.