Attempt at FadeScreen
Implemented a Fade to Black for the Tutorial, Couple of issues: it
flickers, it's the lerp an if statement should fix that. The function
values need to be revised. I call from the update method a function that
uses a newcolor and deltatime to lerp the color. Removed the timer but
it can be readded.